+++ lib/src/app/models/flutter_installer_api/latest_release.model.dart
@@ -1,24 +1,20 @@
-import 'dart:convert';
-
 import 'package:flutter_installer/src/app/models/flutter_installer_api/latest.model.dart';
 
-LatestRelease latestReleaseFromJson(String str) =>
-    LatestRelease.fromJson(json.decode(str));
-
-String latestReleaseToJson(LatestRelease data) => json.encode(data.toJson());
-
 class LatestRelease {
-  LatestRelease({
-    this.latest,
-  });
-
   Latest latest;
 
-  factory LatestRelease.fromJson(Map<String, dynamic> json) => LatestRelease(
-        latest: Latest.fromJson(json["latest"]),
-      );
+  LatestRelease({this.latest});
 
-  Map<String, dynamic> toJson() => {
-        "latest": latest.toJson(),
-      };
+  LatestRelease.fromJson(Map<String, dynamic> json) {
+    latest =
+        json['latest'] != null ? new Latest.fromJson(json['latest']) : null;
+  }
+
+  Map<String, dynamic> toJson() {
+    final Map<String, dynamic> data = new Map<String, dynamic>();
+    if (this.latest != null) {
+      data['latest'] = this.latest.toJson();
+    }
+    return data;
+  }
 }